Create components out of question parts
