+++ packages/webcert/src/feature/certificate/Question/Question.tsx
@@ -1,42 +1,13 @@
-import {
-  CertificateDataConfig,
-  CertificateDataElement,
-  ConfigTypes,
-  Icon,
-  InfoBox,
-  MandatoryIcon,
-  UvText,
-  ValueType,
-  validateExpression,
-} from '@frontend/common'
+import { CertificateDataConfig, ConfigTypes, Icon, MandatoryIcon, UvText } from '@frontend/common'
 import _ from 'lodash'
 import * as React from 'react'
 import { useEffect } from 'react'
 import { useSelector } from 'react-redux'
 import ReactTooltip from 'react-tooltip'
 import { getIsEditable, getIsLocked, getQuestion } from '../../../store/certificate/certificateSelectors'
-import UeCauseOfDeath from '../Inputs/UeCauseOfDeath/UeCauseOfDeath'
-import UeCauseOfDeathList from '../Inputs/UeCauseOfDeath/UeCauseOfDeathList'
-import UeCheckbox from '../Inputs/UeCheckbox'
-import UeCheckboxDateGroup from '../Inputs/UeCheckboxDateGroup'
-import UeCheckboxGroup from '../Inputs/UeCheckboxGroup'
-import UeDate from '../Inputs/UeDate'
-import UeDiagnoses from '../Inputs/UeDiagnosis/UeDiagnoses'
-import UeDropdown from '../Inputs/UeDropdown'
-import UeIcf from '../Inputs/UeIcf'
-import UeMedicalInvestigationList from '../Inputs/UeMedicalInvestigation/UeMedicalInvestigationList'
-import UeMessage from '../Inputs/UeMessage'
-import UeRadio from '../Inputs/UeRadio'
-import UeRadioGroup from '../Inputs/UeRadioGroup'
-import UeRadioGroupOptionalDropdown from '../Inputs/UeRadioGroupOptionalDropdown'
-import { UeSickLeavePeriod } from '../Inputs/UeSickLeavePeriod/UeSickLeavePeriod'
-import UeTextArea from '../Inputs/UeTextArea'
-import UeTextField from '../Inputs/UeTextField'
-import UeTypeahead from '../Inputs/UeTypeahead'
-import UeUncertainDate from '../Inputs/UeUncertainDate'
-import QuestionAccordion from './QuestionAccordion'
 import QuestionHeaderAccordion from './QuestionHeaderAccordion'
 import QuestionHeading from './QuestionHeading'
+import QuestionEditComponent from './QuestionEditComponent'
 
 export interface QuestionProps {
   id: string
@@ -70,6 +41,7 @@ const Question: React.FC<QuestionProps> = ({ id, className }) => {
         </div>
       )
     }
+
     return (
       <>
         {question.config.icon && <Icon iconType={question.config.icon} includeTooltip />}
@@ -79,78 +51,10 @@ const Question: React.FC<QuestionProps> = ({ id, className }) => {
     )
   }
 
-  function getUnifiedEditComponent(question: CertificateDataElement, disabled: boolean) {
-    const commonProps = { key: question.id, disabled, question }
-
-    switch (question.config.type) {
-      case ConfigTypes.UE_RADIO_BOOLEAN:
-        return <UeRadio {...commonProps} />
-      case ConfigTypes.UE_ICF:
-        return <UeIcf {...commonProps} />
-      case ConfigTypes.UE_TEXTAREA:
-        return <UeTextArea {...commonProps} />
-      case ConfigTypes.UE_CHECKBOX_BOOLEAN:
-        return <UeCheckbox {...commonProps} />
-      case ConfigTypes.UE_CHECKBOX_MULTIPLE_CODE:
-        return <UeCheckboxGroup {...commonProps} />
-      case ConfigTypes.UE_DROPDOWN:
-        return <UeDropdown {...commonProps} />
-      case ConfigTypes.UE_RADIO_MULTIPLE_CODE:
-        return <UeRadioGroup {...commonProps} />
-      case ConfigTypes.UE_CHECKBOX_MULTIPLE_DATE:
-        return <UeCheckboxDateGroup {...commonProps} />
-      case ConfigTypes.UE_SICK_LEAVE_PERIOD:
-        return <UeSickLeavePeriod {...commonProps} />
-      case ConfigTypes.UE_DIAGNOSES:
-        return <UeDiagnoses {...commonProps} />
-      case ConfigTypes.UE_RADIO_MULTIPLE_CODE_OPTIONAL_DROPDOWN:
-        return <UeRadioGroupOptionalDropdown {...commonProps} />
-      case ConfigTypes.UE_UNCERTAIN_DATE:
-        return <UeUncertainDate {...commonProps} />
-      case ConfigTypes.UE_MESSAGE:
-        return <UeMessage {...commonProps} />
-      case ConfigTypes.UE_TYPE_AHEAD:
-        return <UeTypeahead {...commonProps} />
-      case ConfigTypes.UE_TEXTFIELD:
-        return <UeTextField {...commonProps} />
-      case ConfigTypes.UE_DATE:
-        return <UeDate {...commonProps} />
-      case ConfigTypes.UE_CAUSE_OF_DEATH:
-        return <UeCauseOfDeath {...commonProps} />
-      case ConfigTypes.UE_CAUSE_OF_DEATH_LIST:
-        return <UeCauseOfDeathList {...commonProps} />
-      case ConfigTypes.UE_MEDICAL_INVESTIGATION:
-        return <UeMedicalInvestigationList {...commonProps} />
-      case ConfigTypes.UE_HEADER:
-        return
-      default:
-        return <InfoBox type="error">Cannot find a component for: {question.config.type}</InfoBox>
-    }
-  }
-
-  function getEditComponent(question: CertificateDataElement, disabled: boolean) {
-    if (question.config.accordion)
-      return (
-        <div id={question.id}>
-          <QuestionAccordion
-            accordion={question.config.accordion}
-            icon={question.config.icon}
-            open={question.value != null ? validateExpression(`'${question.value.id}'`, question.value as ValueType) : false}>
-            {getUnifiedEditComponent(question, disabled)}
-          </QuestionAccordion>
-        </div>
-      )
-    else return getUnifiedEditComponent(question, disabled)
-  }
-
-  function getUnifiedViewComponent(question: CertificateDataElement) {
-    return <UvText question={question} />
-  }
-
   return (
     <div className={className}>
       {getQuestionComponent(question.config, displayMandatory, question.readOnly)}
-      <div>{question.readOnly ? getUnifiedViewComponent(question) : getEditComponent(question, disabled)}</div>
+      <div>{question.readOnly ? <UvText question={question} /> : <QuestionEditComponent question={question} disabled={disabled} />}</div>
     </div>
   )
 }
